Adding of a netcatks banner, improving netcatks with capability to create dprotocol structures via command line, including an API for it, now its e possible to make a network application in few minutes, adding of a to_tdo method inside DProtocol, soon we will supporting TDO defined by the users, extended a DProtocol subscriber now we cath a case when tuple comming not IValidatorResponse, adding of colorama inside setup.py it's user by netcatks, also the tdo module was removed from DProtocol, the to_tdo method now lives inside DProtocol
